graphql/requirements.txt
ariadne==0.13.0
asgiref==3.4.1
click==8.0.3
graphql-core==3.1.6
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
Werkzeug==2.0.2
